revert pool of connections PR #97
