+++ doc/src/site/sphinx/First_Steps.rst
@@ -93,7 +93,7 @@ Configuration parameters
 +-----------------------------------------------+--------------------------------------------------------------------------------+-------------------------+
 | writeConcern                                  | "safe"                                                                         | No                      |
 +-----------------------------------------------+--------------------------------------------------------------------------------+-------------------------+
-| splitSize                                     |       10                                                                       | No                      |
+| splitSize                                     |      "10"                                                                      | No                      |
 +-----------------------------------------------+--------------------------------------------------------------------------------+-------------------------+
 | splitKey                                      | "fieldName"                                                                    | No                      |
 +-----------------------------------------------+--------------------------------------------------------------------------------+-------------------------+
