--- spark-mongodb/src/main/scala/com/stratio/datasource/mongodb/client/MongodbConnection.scala
@@ -1,27 +1,0 @@
-/**
- * Copyright (C) 2015 Stratio (http://stratio.com)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *         http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.stratio.datasource.mongodb.client
-
-import com.stratio.datasource.mongodb.client.ConnectionStatus._
-import com.stratio.datasource.mongodb.client.MongodbClientFactory.Client
-
-/**
- * Connection object that represents one client connection with the status and the expire session time
- */
-case class MongodbConnection(client: Client,
-                             timeOut: Long = 0L,
-                             status: ConnectionStatus = ConnectionStatus.Busy,
-                             index: Int = 0)
