+++ spark-mongodb/src/test/scala/com/stratio/datasource/mongodb/partitioner/MongodbPartitionerIT.scala
@@ -19,12 +19,11 @@ import com.mongodb.DBObject
 import com.mongodb.util.JSON
 import com.stratio.datasource.MongodbTestConstants
 import com.stratio.datasource.mongodb._
-import com.stratio.datasource.mongodb.client.MongodbClientFactory
-import com.stratio.datasource.mongodb.config.{MongodbConfig, MongodbConfigBuilder}
 import com.stratio.datasource.mongodb.config.MongodbConfig._
+import com.stratio.datasource.mongodb.config.{MongodbConfig, MongodbConfigBuilder}
 import org.junit.runner.RunWith
 import org.scalatest.junit.JUnitRunner
-import org.scalatest.{BeforeAndAfterAll, BeforeAndAfter, Matchers, FlatSpec}
+import org.scalatest.{BeforeAndAfter, BeforeAndAfterAll, FlatSpec, Matchers}
 
 @RunWith(classOf[JUnitRunner])
 class MongodbPartitionerIT extends FlatSpec
@@ -86,9 +85,6 @@ with BeforeAndAfterAll {
     }.map(_._2)
   }
 
-  override def afterAll {
-    MongodbClientFactory.closeAll(false)
-  }
 
   it should "get proper partition ranges using splitVector with bounds" + scalaBinaryVersion in {
